requirements.txt
# This file lists all the necessary packages for the project. To set up the
# virtual environment and install these packages, follow these steps:
# 1. Install virtualenv if it's not already installed:
#    pip install virtualenv
# 2. Create a new virtual environment:
#    python -m virtualenv .venv
# 3. Activate the virtual environment:
#    - On Windows: .\.venv\Scripts\activate
#    - On Windows (powershell): .\.venv\Scripts\activate.ps1
#      ! Try: Run as Administrator and `Set-ExecutionPolicy RemoteSigned`
#    - On macOS/Linux: source .venv/bin/activate
# 4. Install the packages from this file:
#    pip install -r requirements.txt
# 5. When done, deactivate the virtual environment with: deactivate

pyautogui
Pillow
gspread
oauth2client
keyboard
numpy
opencv-python
pytesseract
pynput
pytest
fuzzywuzzy
python-Levenshtein
PyQt5
pywin32
dependencies/tesserocr-2.6.0-cp311-cp311-win_amd64.whl
